Added big update
After I updated the db table names I also had to update the names here. I also added a new module for sanitizing the user input.

The host is now the only person who can:
    -  update and delete an event
    - create, update, and delete surveys

If there is a lot of data it will probably take too long for the request to complete. So there is definetely improve ment to do.